+++ packages/brick_graphql_generators/lib/src/graphql_deserialize.dart
@@ -2,7 +2,7 @@ import 'package:analyzer/dart/element/element.dart';
 import 'package:brick_graphql/graphql.dart';
 import 'package:brick_graphql_generators/src/graphql_fields.dart';
 import 'package:brick_graphql_generators/src/graphql_serdes_generator.dart';
-import 'package:brick_rest_generators/generators.dart' show JsonDeserialize;
+import 'package:brick_json_generators/json_deserialize.dart';
 
 /// Generate a function to produce a [ClassElement] from GraphQL data
 class GraphqlDeserialize extends GraphqlSerdesGenerator