+++ packages/brick_rest_generators/lib/generators.dart
@@ -1,7 +1,3 @@
-export 'package:brick_rest_generators/src/json_serdes_generator.dart';
 export 'package:brick_rest_generators/src/rest_deserialize.dart';
 export 'package:brick_rest_generators/src/rest_serialize.dart';
 export 'package:brick_rest_generators/src/rest_fields.dart';
-
-export 'package:brick_rest_generators/src/json_deserialize.dart';
-export 'package:brick_rest_generators/src/json_serialize.dart';